+++ test/full_test.sol
@@ -1,20 +1,20 @@
 package example
-func sum(values: List<int>) -> long {
-    mut result: long = 0
+func sum(values: List<int>) -> int {
+    mut result: int = 0
     for value in values {
         result = result + value
     }
     return result
 }
-func fibonacci(value: int) -> long {
+func fibonacci(value: int) -> int {
     if value <= 1 {
         return value
     }
-    mut previous: long = 0
-    mut current: long = 1
+    mut previous: int = 0
+    mut current: int = 1
     mut index: int = 2
     while index <= value {
-        next: long = previous + current
+        next: int = previous + current
         previous = current
         current = next
         index = index + 1
@@ -28,11 +28,11 @@ func main() -> int {
         30,
         40
     ]
-    total: long = sum(values)
-    fib: long = fibonacci(20)
+    total: int = sum(values)
+    fib: int = fibonacci(20)
     print("Total: " .. total)
     print("Fibonacci: " .. fib)
-    expected: long = 100
+    expected: int = 100
     if total != expected {
         print("Unexpected total")
         return 1
